put component in separate folder - changed font sizing system - changed coloring system - changed name of line height
